azure_functions/requirements.txt
azure-functions>=1.15.0
azure-storage-blob>=12.18.0
azure-identity>=1.14.0
pandas>=2.0.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
